don't parse firmware as an MSE on the n3g
the n3g's firmware is not an MSE package. trying to parse it as one on an n3g restore will cause it to not accept the firmware payload because the reported payload size is different.

(cherry picked from commit 7bbe93dc6935254e535ef9a0520b18a8d7b459cf)
Signed-off-by: Tucker Osman <osmiumusa@gmail.com>